9:26 PM - 4/3/2025
Fixed an error causing user input (in time/weekday columns) to break the highlighting function, causing it to rapidly flash the next lesson's row instead of just highlight it.  It also didn't unhighlight the old one, so that's fixed as well.

There seems to be an issue (though it only happened once so far) where back to back lessons aren't correctly handled when going from current lesson to current lesson (though the next lesson is updating in real-time as it should).  I have no details at the moment, but it took about a full minute after the "initial current lesson" ended for the "new current lesson" to receive the correct formatting.  The program had about 20 chances to update this, but it seems to have only managed on the ~21st update.  I'm almost certain that this has to do with the fetching of lessons or the program's process of finding out the current lesson in session, because the logs reflected as such.

My head is pretty full at the moment so my memory might be failing me here.  I'll try to come back tomorrow and see what happens.