9:26 PM - 4/3/2025
Fixed an error causing user input (in time/weekday columns) to break the highlighting function, causing it to rapidly flash the next lesson's row instead of just highlight it.  It also didn't unhighlight the old one, so that's fixed as well.

There seems to be a new reoccurring issue where, when several lessons occur back-to-back, it takes the program exactly 1 minute to catch onto the fact that a new lesson is in session.  This happened twice with this exact amount of time (and the same number of updates - exactly 12.  This is referring to highlight_next_and_current.py, the time.sleep line at the end of the while True loop, which I'll just refer to as the reinit delay variable).

I tested this a few times, and changed the reinit delay to 2.5, which got rid of the problem completely.  I tested this twice and got the same result both times.

If it's something I could help, I'd prefer a higher reinit delay and have most of the reinitialization be handled by detected updates to the spreadsheet to minimize overhead, but my experience is limited at the moment so I just want to stick with the reinit delay since it's simpler.