4/4/2025 - 10:37 PM
Successfully mitigated desynchronization related risks and fixed a "one minute late" error whendetermining in-session lessons, specifically with back-to-back lessons.
New issue is the program just runs through the entire spreadsheet once, and thinks that there will never be a guitar lesson ever again once the last lesson on the list is over.